review squads demo
