+++ .env.example
@@ -1,0 +1,1 @@
+NTT_MANAGER_PROGRAM_ID="INSERT_NTT_MANAGER_PROGRAM_ID"
