+++ .gitignore
@@ -2,3 +2,5 @@ package-lock.json
 node_modules
 
 .env
+
+multisig-keys.json
