+++ package.json
@@ -1,12 +1,20 @@
 {
-  "devDependencies": {
-    "typescript": "^5.5.4"
-  },
-  "dependencies": {
-    "@project-serum/anchor": "^0.26.0",
-    "@solana/web3.js": "1.70.3",
-    "@sqds/multisig": "^2.1.3",
-    "@wormhole-foundation/sdk": "^0.14.0",
-    "@wormhole-foundation/sdk-solana-ntt": "^0.4.0"
-  }
+	"scripts": {
+		"create-squad": "npx ts-node src/scripts/createSquad.ts",
+		"transfer-ownership": "npx ts-node src/scripts/transferOwnership.ts",
+		"transfer-ownership-mainnet": "npx ts-node src/scripts/transferOwnershipMainnet.ts",
+		"manage-limits": "npx ts-node src/scripts/manageLimits.ts"
+	},
+	"devDependencies": {
+		"typescript": "^5.5.4"
+	},
+	"dependencies": {
+		"@project-serum/anchor": "^0.26.0",
+		"@solana/web3.js": "^1.95.4",
+		"@sqds/multisig": "^2.1.3",
+		"@wormhole-foundation/sdk": "^0.14.0",
+		"@wormhole-foundation/sdk-solana-ntt": "^0.4.0",
+		"bs58": "^6.0.0",
+		"dotenv": "^16.4.5"
+	}
 }
