+++ src/config/keys.json
@@ -1,0 +1,1 @@
+[]
